Move logo files
