+++ src/components/common/Header/index.tsx
@@ -10,7 +10,7 @@ import SafeTokenWidget, { getSafeTokenAddress } from '@/components/common/SafeTo
 import NotificationCenter from '@/components/notification-center/NotificationCenter'
 import { AppRoutes } from '@/config/routes'
 import useChainId from '@/hooks/useChainId'
-import SafeLogo from '@/public/logo.svg'
+import SafeLogo from '@/public/images/logo.svg'
 import Link from 'next/link'
 
 type HeaderProps = {
