+++ src/components/common/QRCode/index.tsx
@@ -15,7 +15,7 @@ const QRCode = ({ value, size }: { value?: string; size: number }): ReactElement
       bgColor={palette.background.paper}
       fgColor={palette.text.primary}
       imageSettings={{
-        src: '/logo-no-text.svg',
+        src: '/images/logo-no-text.svg',
         width: QR_LOGO_SIZE,
         height: QR_LOGO_SIZE,
         excavate: true,
