+++ src/services/onboard.ts
@@ -41,7 +41,7 @@ export const createOnboard = (chainConfigs: ChainInfo[]): OnboardAPI => {
 
     appMetadata: {
       name: 'Safe',
-      icon: '/logo-no-text.svg',
+      icon: '/images/safe-logo-green.png',
       description: 'Please select a wallet to connect to Safe',
       recommendedInjectedWallets: getRecommendedInjectedWallets(),
     },
