+++ src/services/pairing/connector.ts
@@ -4,7 +4,7 @@ import bowser from 'bowser'
 import packageJson from '../../../package.json'
 import { IS_PRODUCTION } from '@/config/constants'
 import ExternalStore from '@/services/ExternalStore'
-import PairingIcon from '@/public/images/safe-mobile-connect.png'
+import PairingIcon from '@/public/images/safe-logo-green.png'
 
 export const PAIRING_MODULE_STORAGE_ID = 'pairingConnector'
